fix(eucstats): downsample picked avatar photos to stop OOM crash
The two eucstats photo pickers (onboarding + profile edit) decoded the
picked image at full resolution via a bare BitmapFactory.decodeStream.
A typical 12-50 MP phone photo becomes a multi-hundred-MB ARGB_8888
bitmap that OOMs the heap or exceeds RecordingCanvas's 100 MB per-bitmap
draw limit ("trying to draw too large bitmap") the moment the crop
dialog renders it -- crashing the app on "some photos" (the large ones).

The nav-marker picker already solved this with a two-pass downsampled
decode. Extract that into a single shared decodeDownsampledBitmap() and
route all three pickers through it so the downsampling can't be skipped
by a new picker again. Also bump the profile-edit crop to 256px output
to match onboarding (was downgrading the avatar to 64px on re-edit).
