+++ app/src/main/java/com/eried/eucplanet/ui/settings/eucstats/OnlineProfileScreen.kt
@@ -46,6 +46,7 @@ import androidx.compose.ui.unit.dp
 import com.eried.eucplanet.R
 import com.eried.eucplanet.data.eucstats.RiderProfile
 import com.eried.eucplanet.ui.navigator.UserMarkerCropDialog
+import com.eried.eucplanet.ui.navigator.decodeDownsampledBitmap
 import com.eried.eucplanet.ui.settings.SettingsViewModel
 import com.eried.eucplanet.ui.theme.appColors
 import com.eried.eucplanet.ui.theme.themedFieldColors
@@ -126,11 +127,10 @@ fun OnlineProfileDialog(
         ActivityResultContracts.GetContent()
     ) { uri ->
         if (uri == null) return@rememberLauncherForActivityResult
-        val bmp = runCatching {
-            val src = context.contentResolver.openInputStream(uri)
-                ?: return@rememberLauncherForActivityResult
-            android.graphics.BitmapFactory.decodeStream(src)
-        }.getOrNull() ?: return@rememberLauncherForActivityResult
+        // Downsample on decode: a full-res gallery photo would OOM / blow the
+        // canvas bitmap limit in the crop dialog and crash the app.
+        val bmp = runCatching { decodeDownsampledBitmap(context, uri) }
+            .getOrNull() ?: return@rememberLauncherForActivityResult
         pickedBitmap  = bmp
         croppedBitmap = null
         showCropDialog = true
@@ -145,7 +145,9 @@ fun OnlineProfileDialog(
                 onApply  = { cropped ->
                     croppedBitmap  = cropped
                     showCropDialog = false
-                }
+                },
+                // 256px so the preview is crisp (server still re-encodes to 64).
+                outputSize = 256,
             )
             return
         }
